Move some home elements into shared
