+++ lib/modules/story_info/screens/story_info.dart
@@ -1,7 +1,7 @@
 import 'package:andax/models/story.dart';
 import 'package:andax/modules/store_play/screens/play.dart';
 import 'package:andax/modules/story_editor/screens/story_editor.dart';
-import 'package:andax/modules/story_info/services/content_loader.dart';
+import 'package:andax/shared/services/story_loader.dart';
 import 'package:andax/modules/story_translator/screens/crowdsourcing.dart';
 import 'package:andax/shared/widgets/loading_dialog.dart';
 import 'package:andax/shared/widgets/rounded_back_button.dart';
@@ -9,7 +9,7 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 
-import '../services/likes.dart';
+import '../../../shared/services/likes.dart';
 
 class StoryScreen extends StatefulWidget {
   const StoryScreen(
@@ -46,7 +46,7 @@ class _StoryScreenState extends State<StoryScreen> {
     await FirebaseFirestore.instance
         .doc('stories/${i.storyID}/translations/${i.translationID}')
         .update({'metaData.views': FieldValue.increment(1)});
-    await loadExperience(
+    await loadStory(
       context,
       i,
       (s, t) => Navigator.push<void>(
@@ -147,7 +147,7 @@ class _StoryScreenState extends State<StoryScreen> {
                         ),
                         const SizedBox(width: 8),
                         IconButton(
-                          onPressed: () => loadExperience(
+                          onPressed: () => loadStory(
                             context,
                             widget.info,
                             (s, t) => Navigator.push<void>(
